isolate sign function
